export coordinate metadata
This change allows export of the "coordinate" node metadata property.
This value is exported as a tuple of [x,y,z].

If the network view is exported then the node positions will be
recalculated using the network view.

If only the network is exported then the "coordinate" node metadata
property is removed for all nodes.

Refactored coordinate translation into fine-grained functional method
for both absolute and relative coordinate spaces.